added travis ci and added non root user to docker iamge

+++ build.sh
@@ -1,4 +1,5 @@
 #! /bin/bash
+echo $CI $TRAVIS
 echo
 echo "Building go application."
 echo
@@ -7,12 +8,15 @@ echo
 echo "Building docker image."
 echo
 docker build -t hotel-api:latest .
-echo
-echo "Running docker image."
-echo
-docker run -it --rm --name hotel-api --publish 8000:8000 hotel-api
-echo
-echo "Deleting the docker image."
-echo
-docker rmi hotel-api
-echo
+
+if [ -z "$TRAVIS" ]; then
+	echo
+	echo "Running docker image."
+	echo
+	docker run -it --rm --name hotel-api --publish 8000:8000 hotel-api
+	echo
+	echo "Deleting the docker image."
+	echo
+	docker rmi hotel-api
+	echo
+fi
